refactor(kfx): extract package lifecycle commands
Move package installation, inspection, removal, and WASM execution into a same-owner module while retaining the established kfx command and helper exports.
